+++ NEWS.md
@@ -4,6 +4,7 @@
 
   * fix bug and add arguments on ``amTimeSeries`` 
   * Add multiple grouping on ``amBoxplot`` 
+  * fix using labelRotation + xlab
   
 ### rAmCharts 2.1.3 (based on [amcharts][amcharts_url] version **3.20.18**)